backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow>=10.2.0
pandas==2.1.3
openpyxl==3.1.2
apscheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
